+++ Engine/EngineFwd.h
@@ -34,20 +34,34 @@ class xml_oarchive;
 }
 }
 
+class AbstractOfxEffectInstance;
 class AppInstance;
 class AppSettings;
 class Bezier;
 class BezierCP;
 class BlockingBackgroundRender;
+class BooleanParam;
 class BufferableObject;
+class ButtonParam;
 class CLArgs;
 class CacheEntryHolder;
 class ChoiceExtraData;
+class ChoiceParam;
+class ColorParam;
 class Curve;
 class DockablePanelI;
+class Double2DParam;
+class Double3DParam;
+class DoubleParam;
 class Effect;
+class FileParam;
 class Format;
+class GlobalOFXTLS;
+class GroupParam;
 class Hash64;
+class Int2DParam;
+class Int3DParam;
+class IntParam;
 class KeyFrame;
 class KnobBool;
 class KnobButton;
@@ -71,24 +85,35 @@ class NodeGraphI;
 class NodeGuiI;
 class NodeSerialization;
 class NodeSettingsPanel;
+class OfxClipInstance;
 class OfxEffectInstance;
 class OfxImage;
+class OutputFileParam;
 class OverlaySupport;
+class PageParam;
+class Param;
+class ParametricParam;
+class PathParam;
 class PluginMemory;
 class ProcessHandler;
 class ProcessInputChannel;
+class QMutex;
 class RectD;
 class RectI;
 class RenderEngine;
 class RenderStats;
+class RichText_Knob;
+class Roto;
 class RotoContext;
 class RotoDrawableItem;
 class RotoItemSerialization;
 class RotoLayer;
 class Settings;
 class StringAnimationManager;
+class StringParam;
 class TimeLine;
 class ViewerInstance;
+struct TextureRect;
 
 namespace Natron {
 class CacheSignalEmitter;
@@ -96,13 +121,16 @@ class EffectInstance;
 class FrameEntry;
 class FrameKey;
 class FrameParams;
+class GenericAccess;
 class Image;
 class ImageComponents;
 class ImageKey;
 class ImageParams;
 class LibraryBinary;
 class Node;
+class OfxHost;
 class OfxImageEffectInstance;
+class OfxOverlayInteract;
 class OfxParamOverlayInteract;
 class OutputEffectInstance;
 class Plugin;
